Extract old rhino/shell run() implementation
It is now in a separate run_old.js file.

Also:
* Add slime.$api.fp.impore.Action
* Harvest tests from wf CLI into jsh.script.cli
* Add untested jsh.shell.world.exit Action
* Rename slime.jrunscript.shell.internal.run.Export to Exports
